+++ src/js/iov/Player.js
@@ -82,7 +82,6 @@ export default class IOVPlayer extends ListenerBaseClass {
     this.iov = iov;
     this.eid = videoJsElementId;
     this.videoId = `clsp-video-${this.iov.config.clientId}`;
-    console.log('creating player', this.id)
 
     this.initializeVideoElement();
 
@@ -509,8 +508,6 @@ export default class IOVPlayer extends ListenerBaseClass {
     if (this.destroyed) {
       return;
     }
-
-    console.log('destroying player', this.id)
 
     this.destroyed = true;